feat: 글 수정 기능 구현

+++ backend/src/main/java/com/duri/domain/post/dto/PostCursor.java
@@ -15,7 +15,7 @@ public class PostCursor {
 
     private LocalDate date;
 
-    private Integer rate;
+    private Double rate;
 
     private String idToken;
